sstable: split two-level iter out of reader.go
